% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_colors.R
\name{ansi_colors}
\alias{ansi_colors}
\alias{ansi_red}
\alias{ansi_green}
\alias{ansi_yellow}
\alias{ansi_blue}
\alias{ansi_magenta}
\alias{ansi_cyan}
\alias{ansi_bold}
\title{Raw ANSI color functions}
\usage{
ansi_red(x)

ansi_green(x)

ansi_yellow(x)

ansi_blue(x)

ansi_magenta(x)

ansi_cyan(x)

ansi_bold(x)
}
\arguments{
\item{x}{Text to colorize}
}
\value{
Text with ANSI color codes
}
\description{
Direct ANSI escape code functions. These always apply color codes
regardless of terminal support. Use the safe \verb{col_*} functions instead.
}
\keyword{internal}
