% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dir_path.R
\name{clean_dir_path}
\alias{clean_dir_path}
\title{Clean a Directory Path}
\usage{
clean_dir_path(path, check_exists = TRUE, create_if_missing = FALSE)
}
\arguments{
\item{path}{The path to clean}

\item{check_exists}{Whether to check if the directory exists}

\item{create_if_missing}{Whether to try to create the directory if it doesn't exist}
}
\value{
A normalized path or NULL if invalid
}
\description{
Helper function to ensure a directory path is properly formatted.
}
