% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_determine_allow_custom}
\alias{._qconfig_determine_allow_custom}
\title{Determine Allow Custom Setting}
\usage{
._qconfig_determine_allow_custom(type, allow_custom)
}
\description{
Determine Allow Custom Setting
}
\keyword{internal}
