% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_do_interactive_config}
\alias{._qconfig_do_interactive_config}
\title{Perform Interactive Configuration}
\usage{
._qconfig_do_interactive_config(
  var_name,
  description,
  options,
  allow_skip,
  note,
  package,
  section,
  verbose,
  type,
  allow_custom,
  allow_create_dir,
  resolve_paths,
  name = NULL
)
}
\description{
Perform Interactive Configuration
}
\keyword{internal}
