% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_format_success_message}
\alias{._qconfig_format_success_message}
\title{Format Success Message}
\usage{
._qconfig_format_success_message(var_name, value)
}
\description{
Format Success Message
}
\keyword{internal}
