% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_handle_skip_input}
\alias{._qconfig_handle_skip_input}
\title{Handle Skip Input}
\usage{
._qconfig_handle_skip_input(user_input, allow_skip)
}
\description{
Handle Skip Input
}
\keyword{internal}
