% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_is_special_keyword}
\alias{._qconfig_is_special_keyword}
\title{Check if Value Contains Special Path Keywords}
\usage{
._qconfig_is_special_keyword(value)
}
\description{
Check if Value Contains Special Path Keywords
}
\keyword{internal}
