% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_parse_selection_input}
\alias{._qconfig_parse_selection_input}
\title{Parse Selection Input with Resolution Mode Suffixes}
\usage{
._qconfig_parse_selection_input(input)
}
\description{
Parse Selection Input with Resolution Mode Suffixes
}
\keyword{internal}
