% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_process_path_input}
\alias{._qconfig_process_path_input}
\title{Process Path Input}
\usage{
._qconfig_process_path_input(path_input, allow_create_dir = TRUE)
}
\description{
Process Path Input
}
\keyword{internal}
