% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_resolve_special_path}
\alias{._qconfig_resolve_special_path}
\title{Resolve Special Path Keywords}
\usage{
._qconfig_resolve_special_path(path_string, package = NULL)
}
\description{
Resolve Special Path Keywords
}
\keyword{internal}
