% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{._qconfig_validate_and_normalize_params}
\alias{._qconfig_validate_and_normalize_params}
\title{Validate and Normalize qconfig Parameters}
\usage{
._qconfig_validate_and_normalize_params(
  var_name,
  package,
  section,
  description,
  options,
  allow_skip,
  note,
  arg_only,
  type,
  allow_custom,
  allow_create_dir,
  resolve_paths,
  fn_tmpl,
  fn_local,
  verbose
)
}
\description{
Validate and Normalize qconfig Parameters
}
\keyword{internal}
