% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config_interactive.R
\name{._uci_determine_allow_custom}
\alias{._uci_determine_allow_custom}
\title{Determine Allow Custom Setting}
\usage{
._uci_determine_allow_custom(type, allow_custom)
}
\description{
Determine Allow Custom Setting
}
\keyword{internal}
