% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config_interactive.R
\name{._uci_do_interactive}
\alias{._uci_do_interactive}
\title{Perform Interactive Configuration}
\usage{
._uci_do_interactive(
  var_name,
  description,
  options,
  allow_skip,
  note,
  package,
  section,
  verbose,
  type,
  allow_custom,
  allow_create_dir,
  resolve_paths,
  name = NULL
)
}
\description{
Perform Interactive Configuration
}
\keyword{internal}
