% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config_interactive.R
\name{._uci_get_variables}
\alias{._uci_get_variables}
\title{Get Available Variable References for update_config_interactive}
\usage{
._uci_get_variables(current_config, current_var_name, template_types = list())
}
\arguments{
\item{current_config}{Current configuration list}

\item{current_var_name}{Name of the variable being configured (to exclude from options)}

\item{template_types}{Named list of variable types from template}
}
\value{
Character vector of variable names that can be referenced
}
\description{
Returns a list of available path-type variables that can be used as references.
Excludes the current variable to prevent self-reference.
}
\keyword{internal}
