% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config_interactive.R
\name{._uci_is_keyword}
\alias{._uci_is_keyword}
\title{Check if Value Contains Special Path Keywords}
\usage{
._uci_is_keyword(value)
}
\description{
Check if Value Contains Special Path Keywords
}
\keyword{internal}
