% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config_interactive.R
\name{._uci_process_path}
\alias{._uci_process_path}
\title{Process Path Input}
\usage{
._uci_process_path(path_input, allow_create_dir = TRUE)
}
\description{
Process Path Input
}
\keyword{internal}
