% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config_interactive.R
\name{._uci_validate_params}
\alias{._uci_validate_params}
\title{Validate and Normalize update_config_interactive Parameters}
\usage{
._uci_validate_params(
  var_name,
  package,
  section,
  description,
  options,
  allow_skip,
  note,
  arg_only,
  type,
  allow_custom,
  allow_create_dir,
  resolve_paths,
  name,
  verbose
)
}
\description{
Validate and Normalize update_config_interactive Parameters
}
\keyword{internal}
