% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config_interactive.R
\name{._uci_write_value}
\alias{._uci_write_value}
\title{Write Configuration Value}
\usage{
._uci_write_value(
  var_name,
  value,
  package,
  section,
  verbose,
  type = NULL,
  name = NULL
)
}
\description{
Write Configuration Value
}
\keyword{internal}
