% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template_interactive.R
\name{._update_template_copy_to_section_interactive}
\alias{._update_template_copy_to_section_interactive}
\title{Copy to Section Interactive}
\usage{
._update_template_copy_to_section_interactive(template_data, verbose = TRUE)
}
\description{
Copy to Section Interactive
}
\keyword{internal}
