% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template_interactive.R
\name{._update_template_remove_inheritance_interactive}
\alias{._update_template_remove_inheritance_interactive}
\title{Remove Inheritance Interactive}
\usage{
._update_template_remove_inheritance_interactive(
  template_data,
  data_sections,
  verbose = TRUE
)
}
\description{
Remove Inheritance Interactive
}
\keyword{internal}
