% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template_interactive.R
\name{._update_template_select_action}
\alias{._update_template_select_action}
\title{Select Template Action}
\usage{
._update_template_select_action()
}
\description{
Select Template Action
}
\keyword{internal}
