% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template_interactive.R
\name{._update_template_update_variable_interactive}
\alias{._update_template_update_variable_interactive}
\title{Update Variable Interactive}
\usage{
._update_template_update_variable_interactive(
  template_data,
  package,
  verbose = TRUE
)
}
\description{
Update Variable Interactive
}
\keyword{internal}
