% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template_interactive.R
\name{._update_template_view_inheritance_chain_interactive}
\alias{._update_template_view_inheritance_chain_interactive}
\title{View Inheritance Chain Interactive}
\usage{
._update_template_view_inheritance_chain_interactive(
  template_data,
  data_sections,
  verbose = TRUE
)
}
\description{
View Inheritance Chain Interactive
}
\keyword{internal}
