% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{._update_template_would_create_circular_dependency}
\alias{._update_template_would_create_circular_dependency}
\title{Would Create Circular Dependency}
\usage{
._update_template_would_create_circular_dependency(inherit_map, child, parent)
}
\description{
Would Create Circular Dependency
}
\keyword{internal}
