% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.add_section_interactive}
\alias{.add_section_interactive}
\title{Add Section Interactive}
\usage{
.add_section_interactive(template_data, verbose = TRUE)
}
\description{
Add Section Interactive
}
\keyword{internal}
