% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.add_variable_interactive}
\alias{.add_variable_interactive}
\title{Add Variable Interactive}
\usage{
.add_variable_interactive(
  template_data,
  package,
  verbose = TRUE,
  debug = FALSE
)
}
\description{
Add Variable Interactive
}
\keyword{internal}
