% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.clear_inheritance_interactive}
\alias{.clear_inheritance_interactive}
\title{Clear All Inheritance Interactive}
\usage{
.clear_inheritance_interactive(template_data, verbose = TRUE)
}
\description{
Clear All Inheritance Interactive
}
\keyword{internal}
