% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.configure_inheritance_interactive}
\alias{.configure_inheritance_interactive}
\title{Configure Section Inheritance Interactive}
\usage{
.configure_inheritance_interactive(template_data, verbose = TRUE)
}
\description{
Configure Section Inheritance Interactive
}
\keyword{internal}
