% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.copy_to_section_interactive}
\alias{.copy_to_section_interactive}
\title{Copy Variables to Section Interactive}
\usage{
.copy_to_section_interactive(template_data, verbose = TRUE)
}
\description{
Copy Variables to Section Interactive
}
\keyword{internal}
