% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{.create_template_backup}
\alias{.create_template_backup}
\title{Create backup of template file}
\usage{
.create_template_backup(template_path)
}
\description{
Create backup of template file
}
\keyword{internal}
