% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-qconfig-input-utils.R
\name{.determine_allow_custom}
\alias{.determine_allow_custom}
\title{Determine Allow Custom Setting}
\usage{
.determine_allow_custom(type, allow_custom)
}
\arguments{
\item{type}{Variable type (e.g., "logical", "path", "character", etc.)}

\item{allow_custom}{User-specified allow_custom parameter}
}
\value{
Logical value indicating whether custom input should be allowed
}
\description{
Internal helper function to determine if custom input should be allowed
based on variable type when allow_custom is NULL.
}
\keyword{internal}
