% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_env_vars.R
\name{.display_env_vars_from_local}
\alias{.display_env_vars_from_local}
\title{Display Current Environment Variables}
\usage{
.display_env_vars_from_local(var_names = NULL)
}
\arguments{
\item{var_names}{Character vector of environment variable names to display}
}
\value{
Invisibly returns NULL
}
\description{
Shows the current values of the MSGM environment variables.
}
\keyword{internal}
