% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconfig.R
\name{.do_interactive_config}
\alias{.do_interactive_config}
\title{Perform Interactive Configuration}
\usage{
.do_interactive_config(
  var_name,
  description,
  options,
  allow_skip,
  note,
  write,
  package,
  user,
  verbose,
  type
)
}
\arguments{
\item{var_name}{Variable name}

\item{description}{Description text}

\item{options}{Option values}

\item{allow_skip}{Allow skip flag}

\item{note}{Note text}

\item{write}{Write destination}

\item{package}{Package name}

\item{user}{User section}

\item{verbose}{Verbose flag}

\item{type}{Variable type}
}
\value{
Raw result value or NULL if skipped
}
\description{
Internal helper function to handle the interactive configuration process.
}
\keyword{internal}
