% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-write_config_yaml.R
\name{.extract_yaml_header}
\alias{.extract_yaml_header}
\title{Extract YAML Header Comments}
\usage{
.extract_yaml_header(file_path)
}
\description{
Extract YAML Header Comments
}
\keyword{internal}
