% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_config.R
\name{.format_raw_default}
\alias{.format_raw_default}
\title{Format raw template default value for display}
\usage{
.format_raw_default(value)
}
\description{
Format raw template default value for display
}
\keyword{internal}
