% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-file_finding_helpers.R
\name{.generate_corresponding_file}
\alias{.generate_corresponding_file}
\title{Generate Corresponding Filename}
\usage{
.generate_corresponding_file(filename, target_type)
}
\arguments{
\item{filename}{Character string with original filename}

\item{target_type}{Character string, either "template" or "local"}
}
\value{
Character string with corresponding filename
}
\description{
Takes a filename and generates the corresponding paired filename.
}
\keyword{internal}
