% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{.get_config_priority}
\alias{.get_config_priority}
\title{Get configuration with priority resolution}
\usage{
.get_config_priority(
  package = get_package_name(),
  section = "default",
  resolved_local_path = NULL,
  resolved_template_path = NULL,
  case_format = "snake_case",
  verbose = FALSE
)
}
\description{
Get configuration with priority resolution
}
\keyword{internal}
