% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_write_utils.R
\name{.get_current_session_vars}
\alias{.get_current_session_vars}
\title{Get current session environment variables for a package}
\usage{
.get_current_session_vars(package, user = "default")
}
\arguments{
\item{package}{Character string with package name}

\item{user}{Character string for user configuration section}
}
\value{
Character vector of variable names that are currently in session environment
}
\description{
Get current session environment variables for a package
}
\keyword{internal}
