% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-gets.R
\name{.get_option}
\alias{.get_option}
\title{Get Option for Variable from Template}
\usage{
.get_option(var_name, package, fn_tmpl = NULL)
}
\arguments{
\item{var_name}{Variable name}

\item{package}{Package name}

\item{fn_tmpl}{Optional path to custom YAML template file (NULL uses default template)}
}
\value{
Character vector with options, or NULL if not found
}
\description{
Get Option for Variable from Template
}
\keyword{internal}
