% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-icy_colors.R
\name{.has_color}
\alias{.has_color}
\title{Check if terminal supports colors}
\usage{
.has_color()
}
\value{
Logical indicating if colors are supported
}
\description{
Check if terminal supports colors
}
\keyword{internal}
