% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_messaging.R
\name{.icy_alert_danger}
\alias{.icy_alert_danger}
\title{Danger alert message}
\usage{
.icy_alert_danger(msg)
}
\arguments{
\item{msg}{Danger message}
}
\description{
Displays a danger/error message with red X mark when terminal supports it.
}
