% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_messaging.R
\name{.icy_alert_info}
\alias{.icy_alert_info}
\title{Information alert message}
\usage{
.icy_alert_info(msg)
}
\arguments{
\item{msg}{Information message}
}
\description{
Displays an information message with blue info symbol when terminal supports it.
}
