% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-icy_messaging.R
\name{.icy_debug}
\alias{.icy_debug}
\title{Debug message}
\usage{
.icy_debug(msg, ...)
}
\arguments{
\item{msg}{Debug message}

\item{...}{Additional arguments passed to underlying functions}
}
\description{
Debug message function that uses contextual formatting when available.
Falls back to simple message output when contextual is not available.
}
