% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_messaging.R
\name{.icy_inform}
\alias{.icy_inform}
\title{Information message}
\usage{
.icy_inform(msg, ...)
}
\arguments{
\item{msg}{Information message}

\item{...}{Additional arguments (currently unused, for compatibility)}
}
\description{
Displays an information message using base R message().
Plain text output, no special formatting.
}
