% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.interactive_template_builder}
\alias{.interactive_template_builder}
\title{Interactive Template Builder}
\usage{
.interactive_template_builder(
  template_data,
  package,
  template_file,
  verbose = TRUE,
  debug = FALSE
)
}
\arguments{
\item{template_data}{Existing template data structure}

\item{package}{Package name}

\item{template_file}{Path to template file}

\item{verbose}{Show messages}

\item{debug}{Show debug information}
}
\value{
Modified template data structure
}
\description{
Main interactive interface for building and editing templates.
}
\keyword{internal}
