% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_local.R
\name{.is_rstudio_available}
\alias{.is_rstudio_available}
\title{Check if RStudio API is Available}
\usage{
.is_rstudio_available()
}
\value{
Logical indicating if RStudio is available
}
\description{
Check if RStudio API is Available
}
\keyword{internal}
