% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.manage_sections_interactive}
\alias{.manage_sections_interactive}
\title{Manage Sections Interactive}
\usage{
.manage_sections_interactive(template_data, package, verbose = TRUE)
}
\description{
Manage Sections Interactive
}
\keyword{internal}
