% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-gets.R
\name{.normalize_type}
\alias{.normalize_type}
\title{Normalize Type Names}
\usage{
.normalize_type(type)
}
\arguments{
\item{type}{Character string with type name}
}
\value{
Normalized type name
}
\description{
Normalizes type names to standard R types.
Converts: boolean/bool -> logical, dir -> path
}
\keyword{internal}
