% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{.pattern}
\alias{.pattern}
\title{Define search patterns based on case format}
\usage{
.pattern(
  package = get_package_name(),
  case_format = "snake_case",
  file = "local"
)
}
\arguments{
\item{case_format}{Character string specifying the case format}

\item{file}{Character string specifying the file type}
}
\description{
Define search patterns based on case format
}
\keyword{internal}
