% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conflicts.R
\name{.remove_from_renviron}
\alias{.remove_from_renviron}
\title{Remove a variable from .Renviron file}
\usage{
.remove_from_renviron(var_name)
}
\arguments{
\item{var_name}{Character string with the variable name to remove}
}
\description{
Remove a variable from .Renviron file
}
\keyword{internal}
