% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.remove_inheritance_interactive}
\alias{.remove_inheritance_interactive}
\title{Remove Inheritance for Section Interactive}
\usage{
.remove_inheritance_interactive(template_data, data_sections, verbose = TRUE)
}
\description{
Remove Inheritance for Section Interactive
}
\keyword{internal}
