% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.remove_section_interactive}
\alias{.remove_section_interactive}
\title{Remove Section Interactive}
\usage{
.remove_section_interactive(template_data, verbose = TRUE)
}
\description{
Remove Section Interactive
}
\keyword{internal}
