% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{.remove_template_variables}
\alias{.remove_template_variables}
\title{Remove variables from template}
\usage{
.remove_template_variables(template_data, var_names, sections, verbose)
}
\description{
Remove variables from template
}
\keyword{internal}
