% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.remove_variable_interactive}
\alias{.remove_variable_interactive}
\title{Remove Variable Interactive}
\usage{
.remove_variable_interactive(template_data, verbose = TRUE)
}
\description{
Remove Variable Interactive
}
\keyword{internal}
