% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.set_inheritance_interactive}
\alias{.set_inheritance_interactive}
\title{Set Inheritance for Section Interactive}
\usage{
.set_inheritance_interactive(template_data, data_sections, verbose = TRUE)
}
\description{
Set Inheritance for Section Interactive
}
\keyword{internal}
