% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_config.R
\name{.show_config_full}
\alias{.show_config_full}
\title{Full display mode for show_config}
\usage{
.show_config_full(
  package,
  status_df,
  renviron_config,
  local_config,
  section,
  fn_tmpl = NULL,
  fn_local = NULL
)
}
\description{
Full display mode for show_config
}
\keyword{internal}
