% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_config.R
\name{.show_config_standard}
\alias{.show_config_standard}
\title{Standard display mode for show_config}
\usage{
.show_config_standard(package, status_df, display)
}
\description{
Standard display mode for show_config
}
\keyword{internal}
