% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-template_utils.R
\name{.show_next_steps}
\alias{.show_next_steps}
\title{Show Next Steps}
\usage{
.show_next_steps(package, template_path = NULL)
}
\arguments{
\item{package}{Package name}

\item{template_path}{Path to created template}
}
\description{
Displays helpful next steps after template creation.
}
\keyword{internal}
