% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-template_common.R
\name{.streamlined_variable_collection}
\alias{.streamlined_variable_collection}
\title{Streamlined Variable Collection for Create Mode}
\usage{
.streamlined_variable_collection(
  template_data,
  package,
  template_path,
  mode = "create",
  verbose = TRUE
)
}
\arguments{
\item{template_data}{Initial template data structure}

\item{package}{Package name}

\item{template_path}{Path to template file for auto-save}

\item{verbose}{Show messages}
}
\value{
Modified template data structure
}
\description{
Simplified interactive variable collection specifically for create_template().
Handles the streamlined "Add another variable? (Y/n)" flow with auto-save.
}
\keyword{internal}
