% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-pattern.R
\name{.swap_filename}
\alias{.swap_filename}
\title{Swap Config/Template Filename}
\usage{
.swap_filename(filename)
}
\arguments{
\item{filename}{Character string with a config or template filename.}
}
\value{
Character string with the swapped filename.
}
\description{
Converts a config filename to its template counterpart or vice versa.
}
\keyword{internal}
