% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.update_variable_interactive}
\alias{.update_variable_interactive}
\title{Update Variable Interactive}
\usage{
.update_variable_interactive(template_data, package, verbose = TRUE)
}
\description{
Update Variable Interactive
}
\keyword{internal}
