% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_config_yaml.R
\name{.update_yaml_root}
\alias{.update_yaml_root}
\title{Update YAML Root Level with NULL Preservation}
\usage{
.update_yaml_root(config_data, var_list)
}
\description{
Update YAML Root Level with NULL Preservation
}
\keyword{internal}
