% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-write_config_yaml.R
\name{.update_yaml_section}
\alias{.update_yaml_section}
\title{Update YAML Section with NULL Preservation}
\usage{
.update_yaml_section(config_data, var_list, section)
}
\description{
Update YAML Section with NULL Preservation
}
\keyword{internal}
