% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config_file.R
\name{.validate_config_specific}
\alias{.validate_config_specific}
\title{Validate Config-Specific Requirements}
\usage{
.validate_config_specific(config_data, template_data, verbose)
}
\description{
Validate Config-Specific Requirements
}
\keyword{internal}
