% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config_file.R
\name{.validate_local_specific}
\alias{.validate_local_specific}
\title{Validate Local-Specific Requirements}
\usage{
.validate_local_specific(config_data, template_data, verbose)
}
\description{
Validate Local-Specific Requirements
}
\keyword{internal}
