% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config_file.R
\name{.validate_template_specific}
\alias{.validate_template_specific}
\title{Validate Template-Specific Requirements}
\usage{
.validate_template_specific(config_data, package, verbose)
}
\description{
Validate Template-Specific Requirements
}
\keyword{internal}
