% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.view_inheritance_chain_interactive}
\alias{.view_inheritance_chain_interactive}
\title{View Inheritance Chain Interactive}
\usage{
.view_inheritance_chain_interactive(
  template_data,
  data_sections,
  verbose = TRUE
)
}
\description{
View Inheritance Chain Interactive
}
\keyword{internal}
