% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-update_template.R
\name{.would_create_circular_dependency}
\alias{.would_create_circular_dependency}
\title{Check if Setting Inheritance Would Create Circular Dependency}
\usage{
.would_create_circular_dependency(inherit_map, child, parent)
}
\description{
Check if Setting Inheritance Would Create Circular Dependency
}
\keyword{internal}
