% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_package_name.R
\name{get_current_package_name}
\alias{get_current_package_name}
\title{Get the name of the currently executing package}
\usage{
get_current_package_name()
}
\value{
Character string with the package name, or NULL if it cannot be determined.
}
\description{
Determines the name of the package from which the current function is being called.
}
\keyword{internal}
