% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_dir.R
\name{get_package_dir}
\alias{get_package_dir}
\title{Define a function to get the package directory}
\usage{
get_package_dir(package)
}
\description{
The function is designed to provide the path to the git clone to developer
and to the installation path (similar to \code{fs::path_package()} or
\code{system.file()})
}
