% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_renviron_path.R
\name{get_renviron_path}
\alias{get_renviron_path}
\title{Get Path to User's .Renviron File}
\usage{
get_renviron_path()
}
\value{
Character string with the full path to the .Renviron file.
}
\description{
Determines the correct path to the user's .Renviron file based on the operating system.
}
