% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_messaging.R
\name{icy_alert_success}
\alias{icy_alert_success}
\title{Success alert message}
\usage{
icy_alert_success(msg)
}
\arguments{
\item{msg}{Success message}
}
\description{
Displays a success message with green checkmark when terminal supports it.
}
