% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_messaging.R
\name{icy_alert_warning}
\alias{icy_alert_warning}
\title{Warning alert message}
\usage{
icy_alert_warning(msg)
}
\arguments{
\item{msg}{Warning message}
}
\description{
Displays a warning message with yellow exclamation mark when terminal supports it.
}
