% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_colors.R
\name{icy_colors}
\alias{icy_colors}
\title{ANSI Color Support for icy Package}
\description{
Lightweight ANSI color functions with terminal detection.
No external dependencies - uses direct ANSI escape codes.
}
