% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_messaging.R
\name{icy_text}
\alias{icy_text}
\title{Plain text output}
\usage{
icy_text(msg, ...)
}
\arguments{
\item{msg}{Text to output}

\item{...}{Additional arguments passed to cat()}
}
\description{
Outputs plain text using cat(). No special formatting or colors.
}
