% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_colors.R
\name{safe_colors}
\alias{safe_colors}
\alias{col_red}
\alias{col_green}
\alias{col_yellow}
\alias{col_blue}
\alias{col_magenta}
\alias{col_cyan}
\alias{col_bold}
\title{Safe color functions with terminal detection}
\usage{
col_red(x)

col_green(x)

col_yellow(x)

col_blue(x)

col_magenta(x)

col_cyan(x)

col_bold(x)
}
\arguments{
\item{x}{Text to colorize}
}
\value{
Colored text if terminal supports it, plain text otherwise
}
\description{
Color functions that automatically detect terminal support and only
apply colors when appropriate. Safe to use in all contexts.
}
