% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{._update_template_add_variable_interactive}
\alias{._update_template_add_variable_interactive}
\title{Add Variable Interactive}
\usage{
._update_template_add_variable_interactive(
  template_data,
  package,
  verbose = TRUE,
  debug = FALSE
)
}
\description{
Add Variable Interactive
}
\keyword{internal}
