% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{._update_template_clear_inheritance_interactive}
\alias{._update_template_clear_inheritance_interactive}
\title{Clear Inheritance Interactive}
\usage{
._update_template_clear_inheritance_interactive(template_data, verbose = TRUE)
}
\description{
Clear Inheritance Interactive
}
\keyword{internal}
