% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{._update_template_manage_sections_interactive}
\alias{._update_template_manage_sections_interactive}
\title{Manage Sections Interactive}
\usage{
._update_template_manage_sections_interactive(
  template_data,
  package,
  verbose = TRUE
)
}
\description{
Manage Sections Interactive
}
\keyword{internal}
