% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{._update_template_remove_section_interactive}
\alias{._update_template_remove_section_interactive}
\title{Remove Section Interactive}
\usage{
._update_template_remove_section_interactive(template_data, verbose = TRUE)
}
\description{
Remove Section Interactive
}
\keyword{internal}
