% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{._update_template_remove_variable_interactive}
\alias{._update_template_remove_variable_interactive}
\title{Remove Variable Interactive}
\usage{
._update_template_remove_variable_interactive(template_data, verbose = TRUE)
}
\description{
Remove Variable Interactive
}
\keyword{internal}
