% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{._update_template_set_inheritance_interactive}
\alias{._update_template_set_inheritance_interactive}
\title{Set Inheritance Interactive}
\usage{
._update_template_set_inheritance_interactive(
  template_data,
  data_sections,
  verbose = TRUE
)
}
\description{
Set Inheritance Interactive
}
\keyword{internal}
