% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{.get_config_local}
\alias{.get_config_local}
\title{Get configuration from local file}
\usage{
.get_config_local(
  package = get_package_name(),
  user = "default",
  yaml_file = NULL,
  case_format = "snake_case",
  verbose = FALSE
)
}
\description{
Get configuration from local file
}
\keyword{internal}
