% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{.get_config_priority}
\alias{.get_config_priority}
\title{Get configuration with priority resolution}
\usage{
.get_config_priority(
  package = NULL,
  user = "default",
  yaml_file = NULL,
  case_format = "snake_case"
)
}
\description{
Get configuration with priority resolution
}
\keyword{internal}
