% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{.get_config_template}
\alias{.get_config_template}
\title{Get configuration from template file}
\usage{
.get_config_template(
  package = NULL,
  user = "default",
  yaml_file = NULL,
  case_format = "snake_case"
)
}
\description{
Get configuration from template file
}
\keyword{internal}
