% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{.get_config_template}
\alias{.get_config_template}
\title{Get configuration from template file}
\usage{
.get_config_template(
  package = get_package_name(),
  section = "default",
  resolved_template_path = NULL,
  case_format = "snake_case",
  verbose = FALSE
)
}
\description{
Get configuration from template file
}
\keyword{internal}
