% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-gets.R
\name{.get_description}
\alias{.get_description}
\title{Get Description for Variable from Template}
\usage{
.get_description(var_name, package)
}
\arguments{
\item{var_name}{Variable name}

\item{package}{Package name}
}
\value{
Character string with full description, or NULL if not found
}
\description{
Get Description for Variable from Template
}
\keyword{internal}
