% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icy_messaging.R
\name{.icy_bullets}
\alias{.icy_bullets}
\title{Bulleted list output}
\usage{
.icy_bullets(items)
}
\arguments{
\item{items}{Named vector of items to display}
}
\description{
Simple implementation of bulleted list. Takes a named vector
and outputs with basic bullet points.
}
