% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_local.R
\name{.is_command_available}
\alias{.is_command_available}
\title{Check if System Command is Available}
\usage{
.is_command_available(command)
}
\arguments{
\item{command}{Command name to check}
}
\value{
Logical indicating if command exists
}
\description{
Check if System Command is Available
}
\keyword{internal}
