% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{.pattern}
\alias{.pattern}
\title{Define search patterns based on case format}
\usage{
.pattern(package, case_format = "snake_case", file = "local", yml = FALSE)
}
\arguments{
\item{case_format}{Character string specifying the case format}

\item{file}{Character string specifying the file type}
}
\description{
Define search patterns based on case format
}
\keyword{internal}
