% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_var_names.R
\name{get_env_var_names}
\alias{get_env_var_names}
\title{Get Environment Variable Names}
\usage{
get_env_var_names(yaml_file = NULL, package = NULL, case_format = "snake_case")
}
\arguments{
\item{yaml_file}{Character string with the name of the YAML file. If NULL,
the function will search for a file using \code{get_env_vars_yaml}.}

\item{package}{Character string with the package name. If NULL (default),
the function will attempt to determine the current package name automatically.}

\item{case_format}{Character string indicating the case format to use for
searching the YAML file if \code{yaml_file} is NULL. Options are:
"snake_case" (default), "camelCase", "PascalCase", "kebab-case".}
}
\value{
Character vector of environment variable names.
}
\description{
Reads the list of environment variable names from the YAML config file.
If the YAML file name is not provided, it will search for a file using the
\code{get_env_vars_yaml} function with the specified case format.
}
