% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Check if a value is NULL and provide a default}
\usage{
x \%||\% default
}
\arguments{
\item{x}{The value to check}

\item{default}{The default value to use if x is NULL}
}
\value{
x if not NULL, otherwise default
}
\description{
Check if a value is NULL and provide a default
}
\keyword{internal}
